Add getObjectFromPath helper method
And use it to verify paths for the getattr method. Previously getattr
was always returning successfully so even though the list directory
showed only particular entries, you could change directory into anything
name you like.

We fix this by introducing a function which validates a path against the
Alembic file object hierarchy. It is a bit clunky at the moment, we do a
boost string split and then keep on trying to grab child objects
corresponding to the path segments until we run out or hit something
invalid.

It is up to the function that is using the returned object to check that
it is valid.