Re-write to use a "ClassifiedObject"
We extend the path-to-object method to return a classified object which
when possible returns a valid object and some idea of whether we're
looking it it or its properties.

If we're looking at a particular property then the name of that property
(or rather subpath components) will be in the "remainder" string vector
in the ClassifiedObject. In future another lookup that works with that
remainder might be needed if there are complex nestings of properties.

I'm not pretending this is performant, just that it simplifies things
for the moment.