Better support for properties
We add "getPropertyDataFromPath" which works with the remainder returned
by the object look up and returns a PropertyData object that contains
the PropertyHeader information for the targetted property plus its
parent compound.

We use this to return more specific information in the getattr call,
most notably that compound properties are marked as directories.

We also improve readdir to return the child properties of a compound
property.

Finally, we provide a basic implementation of "read" which can handle
single value booleans scalar properties and multi-value float64
scalar properties as they are two types from the example I'm working
with. The implementation is pretty naive but it has been a battle to get
this far as the alembic docs seem to lack substance on direct access of
properties (rather than going through schemas.)

It needs clean up.

It also raises issues about what we're returning as the size of these
properties in getattr. It seems if you return a size that is too small
there then your read calls will be truncated appropriately. Doing a full
read in every getattr would be too expensive I imagine as getting the
right size involves an ascii conversion of all of the binary data to
figure out the exact character count (for our ascii view anyway.)