Switch to having "values" file
To open up future possibilities we move away from having the values in a
file named after the scalar or array attribute and instead make the
attribute name a folder and put a "values" file inside that.

This allows us to have other files at the same level. For example,
"min", "max" and "average" might be reasonable.

Best to view this diff with ignore whitespace on.