Remove xattr methods until we understand them better
When we include them, our sessions end up trying to call "lgetxattr"
which I'm not sure how to implement/hook in to so it seems to call some
version out of our control and get a "No such file or directory" error
on our paths.

It is possible that implementing xattrs would be useful down the road
for some meta-data but it isn't necessary for now and remove these gets
rid of the errors.